BGQL-2: добавлены Seeds для книг и жанров; внесены правки в схему GraphQl

+++ database/migrations/2019_08_20_222448_create_books_table.php
@@ -19,7 +19,7 @@ class CreateBooksTable extends Migration
                 ->comment('наименование');
             $table->text('description')
                 ->comment('описание');
-            $table->tinyInteger('year')
+            $table->smallInteger('year')
                 ->comment('год');
             $table->timestamps();
         });
